requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.5.0
attrs==23.1.0
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
config==0.5.1
env==0.1.0
environs==9.5.0
execnet==2.0.2
Faker==20.0.3
googletrans==4.0.0rc1
greenlet==3.1.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
libretranslatepy==2.1.1
lxml==4.9.3
marshmallow==3.20.1
packaging==23.1
playwright==1.51.0
pluggy==1.2.0
pydantic==2.1.1
pydantic_core==2.4.0
pyee==12.1.1
pytest==7.4.0
pytest-base-url==2.1.0
pytest-playwright==0.7.0
pytest-rerunfailures==12.0
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.31.0
rfc3986==1.5.0
rootpath==0.1.0
setuptools==74.0.0
six==1.16.0
sniffio==1.3.0
swagger-coverage==3.3.0
text-unidecode==1.3
translate==3.6.1
typing_extensions==4.7.1
urllib3==2.0.4
